requirements.txt
# --------- pytorch --------- #
torch<=2.5.1
torchvision>=0.15.0
lightning>=2.0.0
torchmetrics>=0.11.4

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
phonemizer      # phonemization of text
typeguard==2.13.3
kaldiio
tensorboard
librosa
Cython
numpy<=1.26.4
einops
inflect
Unidecode
scipy
torchaudio
matplotlib<=3.9.2
pandas
conformer==0.3.2
diffusers # tested using version == 0.34.0
notebook
ipywidgets
huggingface-hub # tested using version == 0.32.2
gdown
wget
seaborn
transformers<=4.43.3
speechbrain==0.5.16 # for pretrained accent encoder
resemblyzer
amfm_decompy
silero_vad
jiwer
pymcd
fastdtw
ppgs
